+++ config/initializers/sidekiq.rb
@@ -1,0 +1,10 @@
+schedule_file = "config/schedule.yml"
+
+if File.exists?(schedule_file) && Sidekiq.server?
+  Sidekiq::Cron::Job.load_from_hash YAML.load_file(schedule_file)
+end
+
+if Rails.env.test?
+  require 'sidekiq/testing'
+  Sidekiq::Testing.fake! # fake is the default mode
+end